Create README.md
Chat-Bot or conversational interfaces as they are also known, present a new way for individuals to interact with computer system. A chatbot allows a user to simply ask questions in the same manner that they would
address a human. The technology at the core of the chatbot is natural language processing (“NLP”). NLP enables computers to understand natural language as humans do. Whether the language is spoken or written, natural language processing uses artificial intelligence to take real-world input, process it, and make sense of it in a way a computer can understand. The main functions of NLP include text classification, text extraction, machine translation and natural language generation.
The future advancements in machine learning and NLP can make the chat-bots more intelligent which can make a conversation with a chat-bot similar to a conversation with humans. This can be used in query resolution for businesses replacing humans.